create create_xgrid_util.c and create_xgrid_util.h add #pragma acc routine seq to mosaic_util
