+++ tools/libfrencutils/Makefile.am
@@ -33,6 +33,8 @@ libfrencutils_a_SOURCES = affinity.c \
                           constant.h \
                           create_xgrid.c \
                           create_xgrid.h \
+                          create_xgrid_util.c \
+                          create_xgrid_util.h \
                           gradient_c2l.c \
                           gradient_c2l.h \
                           interp.c \
