+++ tools/libfrencutils/create_xgrid.h
@@ -26,23 +26,6 @@
 #define MV 50
 /* this value is small compare to earth area */
 
-#pragma acc routine seq
-double poly_ctrlon(const double lon[], const double lat[], int n, double clon);
-#pragma acc routine seq
-double poly_ctrlat(const double lon[], const double lat[], int n);
-double box_ctrlon(double ll_lon, double ll_lat, double ur_lon, double ur_lat, double clon);
-double box_ctrlat(double ll_lon, double ll_lat, double ur_lon, double ur_lat);
-int get_maxxgrid(void);
-void get_grid_area(const int *nlon, const int *nlat, const double *lon, const double *lat, double *area);
-void get_grid_great_circle_area(const int *nlon, const int *nlat, const double *lon, const double *lat, double *area);
-//void get_grid_area_dimensionless(const int *nlon, const int *nlat, const double *lon, const double *lat, double *area);
-void get_grid_area_no_adjust(const int *nlon, const int *nlat, const double *lon, const double *lat, double *area);
-int clip(const double lon_in[], const double lat_in[], int n_in, double ll_lon, double ll_lat,
-	 double ur_lon, double ur_lat, double lon_out[], double lat_out[]);
-void pimod(double x[],int nn);
-int clip_2dx2d(const double lon1_in[], const double lat1_in[], int n1_in,
-	       const double lon2_in[], const double lat2_in[], int n2_in,
-	       double lon_out[], double lat_out[]);
 int create_xgrid_1dx2d_order1(const int *nlon_in, const int *nlat_in, const int *nlon_out, const int *nlat_out, const double *lon_in,
 			      const double *lat_in, const double *lon_out, const double *lat_out,
 			      const double *mask_in, int *i_in, int *j_in, int *i_out,
@@ -67,9 +50,6 @@ int create_xgrid_2dx2d_order2(const int *nlon_in, const int *nlat_in, const int 
 			      const double *lon_in, const double *lat_in, const double *lon_out, const double *lat_out,
 			      const double *mask_in, int *i_in, int *j_in, int *i_out, int *j_out,
 			      double *xgrid_area, double *xgrid_clon, double *xgrid_clat);
-int clip_2dx2d_great_circle(const double x1_in[], const double y1_in[], const double z1_in[], int n1_in,
-			    const double x2_in[], const double y2_in[], const double z2_in [], int n2_in,
-			    double x_out[], double y_out[], double z_out[]);
 int create_xgrid_great_circle(const int *nlon_in, const int *nlat_in, const int *nlon_out, const int *nlat_out,
 			      const double *lon_in, const double *lat_in, const double *lon_out, const double *lat_out,
 			      const double *mask_in, int *i_in, int *j_in, int *i_out, int *j_out,
